requirements.txt
stashapp-tools
pyyaml
sqlalchemy
alembic
imagehash
opencv-python
gradio
pillow
pip-system-certs